requirements.txt
pandas~=2.2.2
requests~=2.32.3
python-dotenv~=1.0.1
python-dateutil~=2.9.0.post0
pytest~=8.3.2